Entropies

Functions for computing entanglement and Rényi entropies from an MPS or reduced density matrix.

Von Neumann & Rényi entropies

ITransverse.vn_entanglement_entropy — Function

Computes the Von Neumann entanglement entropy of an MPS psi at all links (normalizing if necessary), returns a vector of floats containing the VN entropies

ITransverse.renyi_entropies — Function

Given an input spectrum, normalizes it (unless normalize_eigs=false) and builds the corresponding entropies (S0, S05, S1, S2, S4).

ITransverse.gensym_renyi_entropies — Function

Given an input MPS psi, computes the symmetric generalized entropies by diagonalizing RTM. Returns a NamedTuple (; S0, S05, S1, S2, S4).

ITransverse.gen_renyi2 — Function

Returns generalized Renyi 2 entropy from the RTM tau~tr|PHI><PSI|/

Generalised / SVD-based entropies

ITransverse.generalized_vn_entropy_symmetric — Function

Generalized entropy for a symmetric environment (psiL,psiL) Assuming we're in LEFT GENERALIZED canonical form, by default bring the MPS to it By default, normalizes the eigenvalues of the symmetric RTM.

ITransverse.generalized_svd_vn_entropy — Function

We can compute the "SVD" VN entropy by just doing a right (generalized) sweep

Reduced density matrix

ITransverse.diagonalize_rdm — Function

Given an input MPS, computes the spectra of its (normalized) RDM via SVD decompositions - these can be used to calculate the usual entanglement entropies

ITransverse.diagonalize_rtm_symmetric — Function

Diagonalize the symmetric RTM |psi*><psi| by sweeping from one end.

`   
|  |  |  |  |   _____                ____
D--D--D--D--D--|     |   (same λ) --|    |
               | renv|      ≃       |    |
D--D--D--D--D--|_____|            --|____|
|  |  |  |  |

`

direction = :right sweeps N→2, building right environments (gen. right canonical: ortho center at site N)

direction = :left sweeps 1→N-1, building left environments (gen. right canonical form: ortho center at site 1)

Returns a length N-1 vector of eigenvalue vectors, ordered bond 1 … N-1.

ITransverse.rho2 — Function

Given a set of eigenvalues, simply computes sum(λ_i^2)

ITransverse.rtm2_contracted — Function

Given two MPS <psi and |phi>, computes tr(τₜ^2) the trace of the reduced transition matrices |phi><psi| at the various cuts, by contracting left and right MPS (see also below). The normalization must be given via normalization_factor

At a given cut, we can compute tr(τₜ^2) as the contraction (asterisks denote contracted inds, ie. we trace over topmost with lowmost inds)

      *  *  * 
      |  |  |
o--o--o--o--o  |phi>
|  |      
□==□==□==□==□  <psi|
      |  |  |
o--o--o--o--o  |phi>
|  |      
□==□==□==□==□  <psi|
      |  |  | 
      *  *  *

We do this by building the left and right blocks

o--o--
|  |      =  left
□==□==

and

            ==□==□==□  
right         |  |  |
            --o--o--o  

and appropriately contract them. This should basically amount to the contraction

tr(--[LEFT]==[RIGHT]--[LEFT]==[RIGHT]--)