tMPO Construction

Functions for assembling the temporal MPO/MPS networks.

Block types

ITransverse.FwtMPOBlocks — Type

Basic building blocks for the folded tMPS/tMPO, folded tensors of time evolution Rotated 90deg clockwise: (L,R,P,P') => (P',P,L,R)

ITransverse.FoldtMPOBlocks — Type

Basic building blocks for the folded tMPS/tMPO, folded tensors of time evolution already rotated 90deg clockwise - so the physical indices are "temporal" ones.

Forward tMPO

fw_tMPO, fw_tMPS – build a forward temporal MPO / MPS up to time T.

Forward-backward tMPO

ITransverse.fwback_tMPO — Function

Unfolded tMPO with

nbetai initial steps of imaginary time evolution

nfw steps of forward time evolution

(optionally) a mid_op operator insertion

nback steps of backwards time evolution

nbetaf steps of imaginary time evolution

fwback_tMPS \u2013 convenience wrapper that returns the MPS form directly.

Folded tMPO

ITransverse.folded_tMPO — Function

Builds folded tMPO. Of the ts timesites, the first b.tp.nbeta ones are imaginary time ones. Accepted kwargs: foldop (default=Identity, accepts Array or ITensor), verbose(=false), initbeta_only(=true)

ITransverse.folded_tMPO_in — Function

Build folded tMPO with an extra site at the beginning for the initial state, which we specify as init_tensor with a physical index init_physidx. The extra site should already be incorporated in the ts index list (and we check whether the dimensions match), so effectively we're building a tMPO for Nt = length(ts)-1 timesteps

ITransverse.folded_tMPO_ext — Function

Builds a folded tMPO extended by n_ext sites to the top (ie. end) with the tensor b.WWl or b.WWr, depending on whether LR=:left or right. The input time sites ts must be already of the (extended) length.

After rotation 90deg clockwise, should look like (for a :left) | | | | <- p' legs rho0-o--o--o--o--o--o-fold_op | | | | | | <- p legs`

and the other way round for :right. So a :left tMPOext should have `nextmoreplegs thanp'`

folded_tMPS, folded_left_tMPS, folded_right_tMPS also construct folded temporal MPS variants.

Generic constructor

ITransverse.construct_tMPS_tMPO — Function

construct_tMPS_tMPO(ψ_i::MPS, Ut::Vector{MPO}, ϕ_f::MPS)

|ϕ_f⟩    o———o———o
         |   |   |
Ut[end]  U———U———U
         |   |   |
Ut[…]          ⋮
         |   |   |
Ut[2]    U———U———U
         |   |   |  time
Ut[1]    U———U———U    ↑
         |   |   |    |
⟨ψ_i|    o———o———o    |
                      |
        ⟨L|  T  |R⟩

-Given a set of MPS-[MPO(1),MPO(2)...MPO(Nt)]-MPS defining a TN with 3 spatial sites, constructs two boundary tMPS (⟨L| and |R⟩) the tMPO TR with Nt temporal sites for further use in the power method.

Optionally if return_swapped_T also returns TL = swapprime(TR, 0, 1)`.

psi_i and psi_f are assumed to be a valid (space-like) MPS

each element of the vector Ut is assumed to be a valid MPO whereas the links match, respectively,

The (physical) sites need not be correctly primed to link up in the time direction.

The final MPS will be automatically daggered! If you do not want it to be daggered, use the flag dagger_final=false.